% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CovBM.R
\name{coef.covBM}
\alias{coef.covBM}
\title{coef.covBM}
\usage{
\method{coef}{covBM}(object, unconstrained = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{covBM}}, inheriting from \code{\link[nlme]{corStruct}}.}

\item{unconstrained}{A logical value. If TRUE the coefficients are returned in unconstrained form (as used
in the optimization algorithm). If FALSE the coefficients are returned in "natural" form.}

\item{...}{Additional arguments (not used by this method).}
}
\description{
This is a method function that extracts the scale coefficient associated with a
Brownian motion correlation structure object.
}
\examples{
cov1<-covBM(form=~time|group)
coef(cov1)
}

