% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CovFracBM.R
\name{coef.covFracBM}
\alias{coef.covFracBM}
\title{coef.covFracBM}
\usage{
\method{coef}{covFracBM}(object, unconstrained = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{covFracBM}}, inheriting from \code{\link[nlme]{corStruct}}.}

\item{unconstrained}{A logical value. If TRUE the coefficients are returned in unconstrained form (as used
in the optimization algorithm). If FALSE the coefficients are returned in "natural" form.}

\item{...}{Additional arguments (not used by this method).}
}
\description{
This is a method function that extracts the scale coefficient and Hurst parameter
associated with a fractional Brownian motion correlation structure object.
}
\examples{
cov2<-covFracBM(form=~time|group)
coef(cov2)
}

