% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CovIOU.R
\name{coef.covIOU}
\alias{coef.covIOU}
\title{coef.covIOU}
\usage{
\method{coef}{covIOU}(object, unconstrained = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{covIOU}}, inheriting from \code{\link[nlme]{corStruct}}.}

\item{unconstrained}{A logical value. If TRUE the coefficients are returned in unconstrained form (as used
in the optimization algorithm). If FALSE the coefficients are returned in "natural" form.}

\item{...}{Additional arguments (not used by this method).}
}
\description{
This is a method function that extracts the perturbation and Alpha parameters
associated with an integrated Ornstein-Uhlenbeck (IOU) process correlation structure object.
}
\examples{
cov3<-covIOU(form=~time|group)
coef(cov3)
}

