% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CovBM.R
\name{corMatrix.covBM}
\alias{corMatrix.covBM}
\title{corMatrix.covBM}
\usage{
\method{corMatrix}{covBM}(object, covariate = getCovariate(object), ...)
}
\arguments{
\item{object}{An object of class \code{\link{covBM}}, inheriting from \code{\link[nlme]{corStruct}}.}

\item{covariate}{List of covariate vectors, at which values the correlation matrix,
or list of correlation matrices, are to be evaluated, as for \code{\link[nlme]{corMatrix.corStruct}}.}

\item{...}{Additional arguments (not used by this method).}
}
\description{
This method generates a scaled covariance matrix (or list of matrices), for a "covBM" "corStruct" object.
}

