% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CovIOU.R
\name{covIOU}
\alias{covIOU}
\title{covIOU}
\usage{
covIOU(value = c(1, 1), form = ~1)
}
\arguments{
\item{value}{Vector of length 2 providing starting values for optimisation of the perturbation
parameter of integrated Ornstein-Uhlenbeck process relative to residual error variance and
the Alpha parameter, respectively.}

\item{form}{A one-sided formula of the form ~t|g, where t represents a continuous
variable (usually time) and g represents a grouping factor, i.e. with a separate
integrated Ornstein-Uhlenbeck process modelled at each level.}
}
\value{
An object of class "covIOU" and inheriting from "corStruct".
}
\description{
This is a constructor function for the "covIOU" class, representing an integrated Ornstein-Uhlenbeck (IOU)
process component in terms of a continuous variable. The object created is a special type of
\code{\link[nlme]{corStruct}}.
}
\examples{
cov3<-covIOU(form=~time|group)
}

