% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print_summary_covStructBM.R
\name{print.summary.corStructBM}
\alias{print.summary.corStructBM}
\title{print.summary.corStructBM}
\usage{
\method{print}{summary.corStructBM}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.corStructBM", containing information
on fitted stochastic process component.}

\item{...}{Additional arguments (not used for this method).}
}
\description{
print.summary.corStructBM
}

