% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print_summary_covStructBM.R
\name{summary.covBM}
\alias{summary.covBM}
\title{summary.covBM}
\usage{
\method{summary}{covBM}(object, structName = class(object)[1], ...)
}
\arguments{
\item{object}{An object of class "covBM", containing information
on fitted stochastic process component.}

\item{structName}{An optional character string defining the type of correlation
structure associated with object, as for \code{\link[nlme]{summary.corStruct}}. Defaults to class(object)[1].}

\item{...}{Additional arguments (not used for this method).}
}
\description{
summary.covBM
}

